Attempt at multithreading shown in commented out section. The app does not crash when the threads are introduced, but the TextViews containing the amount of cards in each player's hand do not change. Check with instructor tomorrow to see why this occurs.
I also discovered an error in the code that was corrected (the end game toast was awarded to the wrong person).

Further commits/pushes to GitHub should not contain the files for 'Random Number Generator' and any previous pushes to GitHub that contain these files are in error and should not be considered when addressing the code for this app.